refactor: Mudando average de double para BigDecimal para melhorar o handling das casas decimais

+++ src/main/java/dev/valente/desafiovagaitau/dto/StatisticsResponseDTO.java
@@ -2,6 +2,8 @@ package dev.valente.desafiovagaitau.dto;
 
 import lombok.Builder;
 
+import java.math.BigDecimal;
+
 @Builder
-public record StatisticsResponseDTO(double sum, double average, double min, double max, long count) {
+public record StatisticsResponseDTO(double sum, BigDecimal average, double min, double max, long count) {
 }
